UpdateChecker
We are pleased to announce the integration of an update checker within our plugin, enhancing the user experience for server owners. With this latest feature, server owners will receive notifications directly in the console whenever a new version of the plugin becomes available for download. This improvement ensures that our valued users can effortlessly stay up-to-date with the latest enhancements and bug fixes, guaranteeing a seamless and secure experience. We are committed to continually enhancing our plugin to meet the needs of our dedicated community of server owners.